'增加设备分组'

+++ src/main/java/com/dj/iotlite/api/dto/ProductListDto.java
@@ -16,7 +16,7 @@ public class ProductListDto implements BaseDto {
 
     Integer onlineCount;
 
-    Integer deviceCount;
+    Long deviceCount;
 
     String ver;
 }
